feat(robo/cli): new `-id --instance-id` option
This makes use of the new WatcherFile helper object to have a "stable ID" of sorts we can use to reference either internally within the process or externally via this file.

Unfortunately, relying on `process.pid` was not an option due to some quirks of how certain package managers and process management tools wrap and spawn subprocesses. That would've been a cleaner solution tho ;-;